requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
google-generativeai==0.3.2